Move SellType Enum to it's own module
